pyproject.toml
[project]
name = "nlp-take-home-alexis-spiliotopoulos"
version = "0.1.0"
description = "Solution to Moveo.AI take-home assignment for the NLP Engineer position"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "mistralai>=1.9.3",
    "ollama>=0.5.1",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "rich-click>=1.8.9",
    "tqdm>=4.67.1",
    "typer>=0.16.0",
]

[tool.ruff]
line-length = 120
#indent-width = 4